Common lib changes for Xray integration at test level
PROD-5296
A listener is used to hook into the test run close event. At this point the output.xml has been finalized and saved.
The test plan is extracted from a yaml file containing one or more definitions of platform description and corresponding test plan number.
The product is const defined as 'PROD' and this is appended to the number read from the yaml file.
Xray setup is done automatically from init board so all tests are covered.
An exception is raised if no test_sets.yml file is found or the test plan cannot be extracted.
Add P20_RF and S10 Lyra board types.