Modify board architecture to support test rack
Remove multiple inheritance because it adds too much complexity.

Don't open serial ports when initializing them because they may not
be needed for a test.

Increase delay when working with com ports that come from USB on the device.
When opening ports send a couple newlines followed by delay to prevent
errors when sending the first command.

PROD-6633

Signed-off-by: Andrew Hedin <andrew.hedin@ezurio.com>