PPK daemon and client
This commit adds two files to help manage the PPK2 in a test station.

The "PPK daemon" is a Python script that runs to manage one or more PPK2s attached to a test station. The daemon gets its configuration from the station config. If a station contains PPKs, the daemon will initialize the PPK with an initial voltage output and then start a TCP server to allow it to receive commands from a client. The daemon exits at the end of a test run to turn off the PPK2.

The "PPK client" connects to the daemon over TCP and exchanges JSON-RPC to control the PPK (change voltage, turn output on/off, measure current). Persistent connections from the client are not required to keep the PPK running.

PROD-18630