Add function to build permutations of boards
For the interop test station where there are multiple boards of multiple types, we need a way to run tests on all of the combinations of boards. That is, if there is a Zephyr board and a Lyra24 board in the test station and we're running BLE tests, we need both Zephyr and Lyra24 boards to be the advertiser and the scanner in the advertising tests. This new function will return all of the permutations of the multiple boards in the test station so that all possible modes can be tested.

An exception is provided here for when the boards are running the same firmware. So, when board "A" and board "B" have the same firmware image, there's no need to test "A" + "B" and "B" + "A" since they are identical. This reduces the needless extra test executions on most of the test stations that contain identical boards.

PROD-25305