Prevent programming two different images
When there are multiple versions of a board's image in the releases directory, the previous version of this script would program the board twice (old version and new version). The order wasn't necessarily guaranteed and we don't want to waste time with extra programming. This new version will compare version numbers of the images found for a particular `filename` pattern and only program the most recent image.

PROD-25367